requirements.txt
mesop
google-genai
